Changed models.py

+++ backend/app/models.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, String, Float, Boolean, Date
+from sqlalchemy import Column, String, Boolean, Date, Numeric
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
 
-class VendingLocation(Base):  # ← Changed from Location to VendingLocation
-    __tablename__ = "vending_locations"  # ← This matches your actual table name!
-    
+class VendingLocation(Base):
+    __tablename__ = "vending_locations"
+
     id = Column(String(50), primary_key=True, index=True)
     retailer = Column(String(100))
     machine_id = Column(String(50))
@@ -14,8 +14,8 @@ class VendingLocation(Base):  # ← Changed from Location to VendingLocation
     city = Column(String(100))
     state = Column(String(10))
     zip_code = Column(String(20))
-    latitude = Column(Float)
-    longitude = Column(Float) 
+    latitude = Column(Numeric(10, 8))    
+    longitude = Column(Numeric(11, 8)) 
     type = Column(String(50))
     last_verified = Column(Date)
     is_active = Column(Boolean)
